+++ e2e/pages/DeleteAccountPage.ts
